backend/requirements.txt
Django==5.2.6
djangorestframework==3.16.1
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.7.0
django-filter==24.2
Pillow==11.3.0
gunicorn==21.2.0
psycopg2-binary==2.9.9
whitenoise==6.6.0
